changed parameter type

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/entity/Round.java
@@ -48,7 +48,7 @@ public class Round{
         }
     }
 
-    public Role getRole(int playerId){
+    public Role getRole(Long playerId){
         return playerRoles.get(playerId);
     }
